Write sumAll function and algorithm
Take numOne and numTwo and put them in arr array. Then check if either
numOne or numTwo is less than zero, or if it's a non-integer. If so,
return ERROR. Next if else statement checks whether numOne or numTwo
is the larger number, then set i to the smaller number + 1 and while
i is less than bigger number push i to arr and add one to i. Finally,
return the reduced value of arr.