Add logic to factorial function
The factorial function should first initialize arr to an array with
num as the first and only item in the array.
Then a for loop is ran where the variable i is assigned to varible num
minus one, keeps running until i is less than 1, and i--.
Each loop will push i into arr, where i is always one less than the last
item in arr.

Then return the result of using the reduce method on arr where the first
value in arr gets assigned to product and is multiplied by the next
item.