Begin rewrite of palindrome function
Previous function could not handle later tests. Current function is a
work in progress.

First the str parameter is split into array arr. Next, the punctuation
is reduced out of the array and the index becomes a key with the
punctuation character being saved as the value. Then array arr, with the
punctuation removed, is reversed.