Finalize palindrome function rewrite
Deleted the reverse method and wrote own that saves to punctuation.
An empty array is assigned to finalArr that will be built as reversed
original string array. A for loop will iterate over index,
which is assigned to arrLength, reducing index by one each loop, until
index is less than 0. arrLength is the length of arr at initialization.

In the for loop if index is not equal to a punctuation key then unshift
character at the index of arr into finalArr.
Else unshift the punctuation character at the index of punctuation into
finalArr.

At the end of the function compare the original string to the joined
reverse string of finalArr, return boolean result.