Rework function completely to use filters
The previous reworked function was still giving false trues in the test.
This new rework has the function passing all tests.

Make the given string lowercase then split it into arr. Filter through
arr into copyArr, returning the character if not a space. Repeat for
rest of punctuation until all punctuation is removed from copyArr.
Then let reversedStr be a copy of copyArr that is reversed. Finally,
compared the joined variables and return true if palindrome, false if
not palindrome.