Write function to find oldest person in array
First, the date is saved to Date. A forEach method goes through the arr
array, which is the given array, and if the person object doesn't have
a yearOfDeath key it is given one with todays year as it's value.

Next the array is sorted through to have the oldest be put last in the
array. Finally, pop the last object in the array to get the name for the
test.